--- include/kernel/shell.h
@@ -1,6 +1,0 @@
-#ifndef SHELL_H_
-#define SHELL_H_
-
-void start_shell();
-
-#endif
