+++ string.h
@@ -5,5 +5,4 @@ char * strtok(char *str, const char* delim);
 int32_t get_spaces(char * str);
 int strcmp(const char * str1, const char * str2);
 
-
 #endif
